ACA-72 - Further cleanup/renaming
