+++ src/main/java/uk/gov/hmcts/reform/managecase/service/noc/VerifyNoCAnswersService.java
@@ -8,7 +8,6 @@ import uk.gov.hmcts.reform.managecase.domain.NoCRequestDetails;
 import uk.gov.hmcts.reform.managecase.domain.Organisation;
 import uk.gov.hmcts.reform.managecase.domain.OrganisationPolicy;
 import uk.gov.hmcts.reform.managecase.repository.PrdRepository;
-import uk.gov.hmcts.reform.managecase.service.ChallengeQuestionService;
 import uk.gov.hmcts.reform.managecase.service.NoticeOfChangeService;
 
 import javax.validation.ValidationException;
@@ -17,15 +16,15 @@ import javax.validation.ValidationException;
 public class VerifyNoCAnswersService {
 
     private final NoticeOfChangeService noticeOfChangeService;
-    private final ChallengeQuestionService challengeQuestionService;
+    private final ChallengeAnswerValidator challengeAnswerValidator;
     private final PrdRepository prdRepository;
 
     @Autowired
     public VerifyNoCAnswersService(NoticeOfChangeService noticeOfChangeService,
-                                   ChallengeQuestionService challengeQuestionService,
+                                   ChallengeAnswerValidator challengeAnswerValidator,
                                    PrdRepository prdRepository) {
         this.noticeOfChangeService = noticeOfChangeService;
-        this.challengeQuestionService = challengeQuestionService;
+        this.challengeAnswerValidator = challengeAnswerValidator;
         this.prdRepository = prdRepository;
     }
 
@@ -34,7 +33,7 @@ public class VerifyNoCAnswersService {
         NoCRequestDetails noCRequestDetails = noticeOfChangeService.challengeQuestions(caseId);
         SearchResultViewItem caseResult = noCRequestDetails.getSearchResultViewItem();
 
-        String caseRoleId = challengeQuestionService
+        String caseRoleId = challengeAnswerValidator
             .getMatchingCaseRole(noCRequestDetails.getChallengeQuestionsResult(),
                 verifyNoCAnswersRequest.getAnswers(), caseResult);
